update /usr/share/tor/tor-service-defaults-torrc
For compatibility with Tor version 0.2.8.6-1~d80.jessie+1.

/var/run/tor/control was no longer created and Tor throwing a warning.

Aug 03 21:56:34.000 [warn] Permissions on directory /var/run/tor are too permissive.
Aug 03 21:56:34.000 [warn] Before Tor can create a control socket in "/var/run/tor/control", the directory "/var/run/tor" needs to exist, and to be accessible only by the user and group account that is running Tor.  (On some Unix systems, anybody who can list a socket can connect to it, so Tor is being careful.)

Which resulted in a follow up issues:
- control-port-filter-proxy-python no longer working
- whonixcheck error: Tor's Control Port could not be reached!
- broken Tor Browser new identity button

Thanks to freeman, torjunkie and temple for reporting this issue!

https://forums.whonix.org/t/error-tor-bootstrap-result